Refactoring (Extract AtomPage into its own file)
